Refactored service from main()
